# AST: More efficient way to collect referenced source units