@blishko 2024-11-21
Here are the results I measured on my laptop.

Before:
|         File          | Pipeline |   Time   | Memory (peak) | Exit code |
|-----------------------|----------|---------:|--------------:|----------:|
| uniswap-v4-2022-06-16 |   legacy |      0 s |       108 MiB |         1 |
| uniswap-v4-2022-06-16 |       ir |      0 s |       106 MiB |         1 |
| openzeppelin-5.0.2    |   legacy |      4 s |       457 MiB |         0 |
| openzeppelin-5.0.2    |       ir |     13 s |       525 MiB |         0 |
| openzeppelin-4.9.0    |   legacy |      5 s |       550 MiB |         0 |
| openzeppelin-4.9.0    |       ir |     16 s |       616 MiB |         0 |
| liquity-2024-10-30    |   legacy |      5 s |       536 MiB |         0 |
| liquity-2024-10-30    |       ir |     17 s |       747 MiB |         0 |
| openzeppelin-4.7.0    |   legacy |      8 s |       573 MiB |         0 |
| openzeppelin-4.7.0    |       ir |     23 s |       597 MiB |         0 |
| openzeppelin-4.8.0    |   legacy |     10 s |       715 MiB |         0 |
| openzeppelin-4.8.0    |       ir |     26 s |       724 MiB |         0 |
| uniswap-v4-2024-06-06 |   legacy |     49 s |      1649 MiB |         0 |
| uniswap-v4-2024-06-06 |       ir |     49 s |      1825 MiB |         0 |
| eigenlayer-0.3.0      |   legacy |     36 s |      2419 MiB |         0 |
| eigenlayer-0.3.0      |       ir |    205 s |      5218 MiB |         0 |
| sablier-v2-1.2.0      |   legacy |    120 s |      5582 MiB |         0 |
| sablier-v2-1.2.0      |       ir |    580 s |      6834 MiB |         0 |
| seaport-1.6           |   legacy |     75 s |      3677 MiB |         0 |
| seaport-1.6           |       ir |      6 s |       916 MiB |         1 |
| farcaster-3.1.0       |   legacy |     11 s |       866 MiB |         0 |
| farcaster-3.1.0       |       ir |     48 s |      1673 MiB |         1 |

After:

|         File          | Pipeline |   Time   | Memory (peak) | Exit code |
|-----------------------|----------|---------:|--------------:|----------:|
| uniswap-v4-2022-06-16 |   legacy |      0 s |       108 MiB |         1 |
| uniswap-v4-2022-06-16 |       ir |      0 s |       106 MiB |         1 |
| openzeppelin-5.0.2    |   legacy |      5 s |       447 MiB |         0 |
| openzeppelin-5.0.2    |       ir |     14 s |       511 MiB |         0 |
| openzeppelin-4.9.0    |   legacy |      6 s |       525 MiB |         0 |
| openzeppelin-4.9.0    |       ir |     17 s |       597 MiB |         0 |
| liquity-2024-10-30    |   legacy |      5 s |       518 MiB |         0 |
| liquity-2024-10-30    |       ir |     18 s |       758 MiB |         0 |
| openzeppelin-4.7.0    |   legacy |      9 s |       563 MiB |         0 |
| openzeppelin-4.7.0    |       ir |     24 s |       603 MiB |         0 |
| openzeppelin-4.8.0    |   legacy |     10 s |       711 MiB |         0 |
| openzeppelin-4.8.0    |       ir |     27 s |       731 MiB |         0 |
| uniswap-v4-2024-06-06 |   legacy |     49 s |      1853 MiB |         0 |
| uniswap-v4-2024-06-06 |       ir |     50 s |      1695 MiB |         0 |
| eigenlayer-0.3.0      |   legacy |     36 s |      2443 MiB |         0 |
| eigenlayer-0.3.0      |       ir |    207 s |      4487 MiB |         0 |
| sablier-v2-1.2.0      |   legacy |     69 s |      5071 MiB |         0 |
| sablier-v2-1.2.0      |       ir |    543 s |      6608 MiB |         0 |
| seaport-1.6           |   legacy |     74 s |      3536 MiB |         0 |
| seaport-1.6           |       ir |      4 s |       962 MiB |         1 |
| farcaster-3.1.0       |   legacy |     11 s |       860 MiB |         0 |
| farcaster-3.1.0       |       ir |     48 s |      1648 MiB |         1 |


There is a big difference for `sablier-v2-1.2.0` on both `legacy` and `ir`. Other projects do not seem to be affected.