@cameel 2024-11-21
> |         File          | Pipeline |   Time   | Memory (peak) | Exit code |
> |-----------------------|----------|---------:|--------------:|----------:|
> | uniswap-v4-2022-06-16 |   legacy |      0 s |       108 MiB |         1 |
> | uniswap-v4-2022-06-16 |       ir |      0 s |       106 MiB |         1 |

What is the error you're getting there (you can see the content of stderr when the script finishes). For me these two pass without errors.