@blishko 2024-11-21
> What is the error you're getting there (you can see the content of stderr when the script finishes)? For me these two pass without errors.

```
==================================
stderr for uniswap-v4-2022-06-16 via legacy
==================================
Warning: Found unknown config section in foundry.toml: [ci]
This notation for profiles has been deprecated and may result in the profile not being registered in future versions.
Please use [profile.ci] instead or run `forge config --fix`.
Warning: Found unknown config section in foundry.toml: [default]
This notation for profiles has been deprecated and may result in the profile not being registered in future versions.
Please use [profile.default] instead or run `forge config --fix`.
Error: Compiler run failed:
Warning (3805): This is a pre-release compiler version, please do not use it in production.

Error (7792): Function has override specified but does not override anything.
  --> contracts/test/MockContract.sol:35:41:
   |
35 |     function _beforeFallback() internal override {
   |                                         ^^^^^^^^

Warning (3628): This contract has a payable fallback function, but no receive ether function. Consider adding a receive ether function.
 --> contracts/test/MockContract.sol:8:1:
  |
8 | contract MockContract is Proxy {
  | ^ (Relevant source part starts here and spans across multiple lines).
Note: The payable fallback function is defined here.
  --> node_modules/@openzeppelin/contracts/proxy/Proxy.sol:66:5:
   |
66 |     fallback() external payable virtual {
   |     ^ (Relevant source part starts here and spans across multiple lines).
```