@cameel 2024-11-21
> ```
> Error (7792): Function has override specified but does not override anything.
>   --> contracts/test/MockContract.sol:35:41:
> ```

> ```
> Warning (3628): This contract has a payable fallback function, but no receive ether function. > Consider adding a receive ether function.
>  --> contracts/test/MockContract.sol:8:1:
> ```

`MockContract.sol` imports `@openzeppelin/contracts/proxy/Proxy.sol`, which for this older Uniswap version had to be installed via npm. The errors look as if it either does not get installed or a wrong version gets installed. Check what you have in `uniswap-v4-2022-06-16/node_modules/@openzeppelin/contracts/proxy/Proxy.sol`. Maybe `external-setup.sh` needs an adjustment. Or maybe you do not have npm?

Though in the end this particular project is not that relevant here (it was mostly useful for benchmarking across a wide range of solc versions) so it's fine to also ignore it if you don't want to dig into it.