@cameel 2024-11-21
And here are my benchmarks:

|         File          | Pipeline | Time (before) | Time (after) | Memory (before) | Memory (after) |
|-----------------------|----------|--------------:|-------------:|----------------:|---------------:|
| uniswap-v4-2022-06-16 |   legacy |           5 s |          5 s |         245 MiB |        244 MiB |
| uniswap-v4-2022-06-16 |       ir |          19 s |         20 s |         298 MiB |        300 MiB |
| openzeppelin-5.0.2    |   legacy |          10 s |         10 s |         450 MiB |        450 MiB |
| openzeppelin-5.0.2    |       ir |          34 s |         30 s |         492 MiB |        492 MiB |
| openzeppelin-4.9.0    |   legacy |          12 s |         13 s |         516 MiB |        516 MiB |
| openzeppelin-4.9.0    |       ir |          39 s |         37 s |         575 MiB |        574 MiB |
| liquity-2024-10-30    |   legacy |          10 s |         10 s |         506 MiB |        504 MiB |
| liquity-2024-10-30    |       ir |          41 s |         39 s |         640 MiB |        641 MiB |
| openzeppelin-4.7.0    |   legacy |          19 s |         18 s |         605 MiB |        606 MiB |
| openzeppelin-4.7.0    |       ir |          52 s |         51 s |         578 MiB |        579 MiB |
| openzeppelin-4.8.0    |   legacy |          21 s |         21 s |         759 MiB |        761 MiB |
| openzeppelin-4.8.0    |       ir |          61 s |         59 s |         729 MiB |        729 MiB |
| uniswap-v4-2024-06-06 |   legacy |          21 s |         21 s |         896 MiB |        898 MiB |
| uniswap-v4-2024-06-06 |       ir |         111 s |        109 s |        1492 MiB |       1484 MiB |
| eigenlayer-0.3.0      |   legacy |          79 s |         72 s |        2402 MiB |       2405 MiB |
| eigenlayer-0.3.0      |       ir |         502 s |        485 s |        4461 MiB |       4457 MiB |
| sablier-v2-1.2.0      |   legacy |         196 s |        150 s |        6345 MiB |       6350 MiB |
| sablier-v2-1.2.0      |       ir |        1267 s |       1258 s |       14513 MiB |      14546 MiB |

The relative difference for sablier is much smaller but it's still there, so there must be something to it.

Also, the huge discrepancy in memory usage is interesting. Apparently sablier uses 15 GB on my machine, which I'd normally ascribe to the recent bump to 1.2 but for you it's much smaller - why? And the difference for `uniswap-v4-2024-06-06` is also quite big.